feat: 6장 pod identity 작업중

+++ 06_Workload_Access_2/01_app_irsa/05_create_role.sh
@@ -8,6 +8,8 @@ fi
 
 if [ ! -f "./local_env.sh" ];then
 	OIDC_ID=$(aws eks describe-cluster --name ${CLUSTER_NAME} --query "cluster.identity.oidc.issuer" --output text ${PROFILE_STRING} | cut -d '/' -f 5)
+else
+  . ./local_env.sh
 fi
 
 SERVICE_ACCOUNT_NAME=eks-edu-service-account-${IDE_NAME}
@@ -19,6 +21,14 @@ if [ ! -d "tmp" ]; then
     mkdir -p tmp
 fi
 
+# Check if role exists
+ROLE_EXISTS=$(aws iam get-role --role-name ${ROLE_NAME} ${PROFILE_STRING} 2>&1 || echo "ROLE_NOT_FOUND")
+
+if [[ ! "$ROLE_EXISTS" == *"ROLE_NOT_FOUND"* ]]; then
+    echo "Role ${ROLE_NAME} 이 존재합니다."
+    exit 1
+fi
+
 cat >tmp/trust-relationship.json <<EOF
 {
   "Version": "2012-10-17",
@@ -26,7 +36,7 @@ cat >tmp/trust-relationship.json <<EOF
     {
       "Effect": "Allow",
       "Principal": {
-        "Federated": "arn:aws:iam::${AWS_ACCOUNT_ID}:oidc-provider/oidc.eks.${REGION_NAME}.amazonaws.com/id/${OIDC_ID}"
+        "Federated": "arn:aws:iam::${AWS_ACCOUNT_ID}:oidc-provider/oidc.eks.${AWS_REGION}.amazonaws.com/id/${OIDC_ID}"
       },
       "Action": "sts:AssumeRoleWithWebIdentity",
       "Condition": {
@@ -42,12 +52,13 @@ EOF
 
 echo "aws iam create-role \\
     --role-name ${ROLE_NAME} \\
-    --assume-role-policy-document file://trust-relationship.json \\
+    --assume-role-policy-document file://tmp/trust-relationship.json \\
     --description "${ROLE_NAME}-description" ${PROFILE_STRING}"
 
+echo "${ROLE_NAME} 생성중..."
 aws iam create-role \
     --role-name ${ROLE_NAME} \
-    --assume-role-policy-document file://trust-relationship.json \
+    --assume-role-policy-document file://tmp/trust-relationship.json \
     --description "${ROLE_NAME}-description" ${PROFILE_STRING}
 
 echo "aws iam attach-role-policy \\
@@ -57,3 +68,4 @@ echo "aws iam attach-role-policy \\
 aws iam attach-role-policy \
     --role-name ${ROLE_NAME} \
     --policy-arn=arn:aws:iam::${AWS_ACCOUNT_ID}:policy/${POLICY_NAME} ${PROFILE_STRING}
+echo "${ROLE_NAME} 생성 완료..."
